chore: Refactor Auth0ProviderWithNavigate
