+++ src/index.js
@@ -5,43 +5,15 @@ import "antd/dist/antd.min.css";
 import "./redesign/style/App.css";
 import PolicyEngine from "./redesign/components/PolicyEngine.jsx";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration.js";
-import { useNavigate, BrowserRouter as Router } from "react-router-dom";
+import { BrowserRouter as Router } from "react-router-dom";
 import loc_en_gb from "./lang/plotly_locales/locale-en-gb.js";
 import loc_en_us from "./lang/plotly_locales/locale-en-us.js";
-import { Auth0Provider } from "@auth0/auth0-react";
+import Auth0ProviderWithNavigate from "./contexts/Auth0ProviderWithNavigate.jsx";
 
 var Plotly = require("plotly.js/dist/plotly.js");
 Plotly.register(loc_en_gb);
 Plotly.register(loc_en_us);
 
-export default function Auth0ProviderWithNavigate({children}) {
-  const navigate = useNavigate();
-
-  const redirectUri = "http://localhost:3000/callback";
-  const domain = process.env.REACT_APP_AUTH0_DOMAIN;
-  const clientId = process.env.REACT_APP_AUTH0_CLIENT_ID;
-
-  const onRedirectCallback = (appState) => {
-    navigate(appState?.returnTo || window.location.pathname);
-  };
-
-  if (!(domain && clientId && redirectUri)) {
-    return null;
-  }
-
-  return (
-    <Auth0Provider
-      domain={domain}
-      clientId={clientId}
-      authorizationParams={{
-        redirect_uri: redirectUri,
-      }}
-      onRedirectCallback={onRedirectCallback}
-    >
-      {children}
-    </Auth0Provider>
-  );
-}
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
